[🔥 Remove] 식단 기능 삭제로 인한 파일 삭제 (#none)
