+++ types/calendar.ts
@@ -1,0 +1,6 @@
+export interface CalendarDayCell {
+  date: string | null;
+  day: number;
+  isCurrentMonth: boolean;
+  isToday: boolean;
+}
