--- util/diet/index.ts
@@ -1,2 +1,0 @@
-export { buildCalendarGrid } from "./buildCalendarGrid";
-export { buildDietDayMarkers } from "./buildDietDayMarkers";
